restrucre cuda backend quantize op dispatch
