+++ backends/cuda/quantize_op_dispatch/_library.py
@@ -1,0 +1,16 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""Shared torch.library handle for the ``executorch_cuda`` op namespace.
+
+``int4_dispatch`` and ``int8_dispatch`` both register custom ops into the same
+``executorch_cuda`` namespace, so they must share a single ``DEF`` library
+instance — PyTorch allows only one ``DEF`` per namespace per process.
+"""
+
+from torch.library import Library
+
+lib = Library("executorch_cuda", "DEF")
